feat: change address to new format
after another discussion with max, region can be changed to
state, whereas state can be used for administrative regions
as well to accomodate the naming conventions in countries
such as France.
Note that this field may be refactored to iso3166-2
region codes depending on UX research and availabitily of
reliable data sources/libraries.